グローバル変数、メソッドの整理
